move shared to db
