+++ src/cli/actions/settings/emergencyKit.js
@@ -4,8 +4,8 @@ const qrcode = require('qrcode')
 const { PDFDocument, StandardFonts, rgb } = require('pdf-lib')
 const { logger } = require('@dotenvx/dotenvx')
 
-const current = require('./../../../shared/current')
-const userPrivateKey = require('./../../../shared/userPrivateKey')
+const current = require('./../../../db/current')
+const userPrivateKey = require('./../../../db/userPrivateKey')
 const smartMask = require('./../../../lib/helpers/smartMask')
 const maskRecoveryPhrase = require('./../../../lib/helpers/maskRecoveryPhrase')
 const formatRecoveryPhrase = require('./../../../lib/helpers/formatRecoveryPhrase')
