+++ src/db/helpers/extractSubdomainAndDomain.js
